refactor(core): move read postcondition postprocessing
